Harden schema migrations and degraded startup
