+++ scripts/test.sh
@@ -27,6 +27,8 @@ TEST_IMAGE="${MOBIUS_IMAGE:-mobius-test:ci}"
 # suite remains the required --backend/--all closeout gate.
 FAST_TESTS=(
   "tests/test_readiness.py"
+  "tests/test_db_migrations.py::test_previous_release_database_upgrades_to_current_orm"
+  "tests/test_db_migrations.py::test_published_schema_migration_history_is_unique_ordered_and_immutable"
   "tests/test_auth_helpers.py"
   "tests/test_app_compile_contract.py"
   "tests/test_source_status.py"
